+++ R/defaultControl.R
@@ -9,7 +9,8 @@ defaultControl<-function() {
     drainage = TRUE,
     leafPhenology = TRUE,
     transpirationMode = "Granier", #vs. "Sperry"
-    hydraulicCostFunction = 1, #vs. 2
+    gainModifier = 1, 
+    costModifier = 1, 
     verticalLayerSize = 100,
     # nStemSegments = 1,
     capacitance = FALSE,