+++ src/transpiration.h
@@ -5,7 +5,8 @@
 #endif
 using namespace Rcpp;
 
-List profitMaximization(List supplyFunction, DataFrame photosynthesisFunction, int type, double Gwmin, double Gmax, double kleafmax = NA_REAL);
+List profitMaximization(List supplyFunction, DataFrame photosynthesisFunction, double Gwmin, double Gwmax, 
+                        double gainModifier = 1.0, double costModifier = 1.0);
 List transpirationGranier(List x, List soil, double tday, double pet, bool modifyInput = true);
 List transpirationSperry(List x, List soil, double tmin, double tmax, double rhmin, double rhmax, double rad, double wind, 
                   double latitude, double elevation, double slope, double aspect,